### 2021-02-10 5a0bf42
Update check-release.yaml
