+++ .github/workflows/check-release.yaml
@@ -20,7 +20,12 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: r-lib/actions/setup-r@v1
-      - name: Install dependencies
+      - name: Install system dependencies
+        run: |
+          sudo apt-get update -qq
+          sudo apt-get install texlive-latex-base libharfbuzz-dev libfribidi-dev
+          sudo apt-get install gsl-bin libgsl0-dev
+      - name: Install package dependencies
         run: |
           install.packages(c("remotes", "rcmdcheck"))
           remotes::install_deps(dependencies = TRUE)
