+++ .github/workflows/check-release.yaml
@@ -10,11 +10,11 @@ on:
       - main
       - master
 
-name: R-CMD-check
+name: R-CMD-check-release
 
 jobs:
   R-CMD-check:
-    runs-on: macOS-latest
+    runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
