+++ .github/workflows/check-release.yaml
@@ -25,6 +25,7 @@ jobs:
           sudo apt-get update -qq
           sudo apt-get install texlive-latex-base libharfbuzz-dev libfribidi-dev
           sudo apt-get install gsl-bin libgsl0-dev
+          sudo apt-get install libcurl4-openssl-dev
       - name: Install package dependencies
         run: |
           install.packages(c("remotes", "rcmdcheck"))
