### 2021-02-10 e377123
Update check-release.yaml
