### 2021-02-22 #3027 @marcosvital
Hi,  @gmyenni, I will be the editor of your submission.

Please take a look at the errors compiled above regarding the pdf generation, and check if you can correct them. If you make any changes and want to try again, you can use `@whedon generate pdf` to have whedon try to create a new version (or just let me know and I'll do it).

At the same time, we can start assigning reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this - you can paste the github profile of anyone you think that could do it (but please don't tag anyone yet).
You can also check [this list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS, and see if any of them may be suitable to review this submission.
