### 2021-02-26 669001c
Use RStudio package manager in GH Actions
This speeds up the package installation by about 15 minutes by using
RStudio Package manager binaries for Ubuntu.
