+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   current_models: current_models.html
   getting_started: getting_started.html
   jags_models: jags_models.html
-last_built: 2021-02-26T17:31Z
+last_built: 2021-02-26T17:35Z
 urls:
   reference: https://weecology.github.io/portalcasting/reference
   article: https://weecology.github.io/portalcasting/articles
