### 2021-03-02 #3027 @danielskatz
If you can say a little about the COI, we can decide if it can be waived or will prevent you from reviewing.  Of course, if you prefer not to review because of the COI, that's ok too.
