### 2021-03-03 #3027 @ViniciusBRodrigues
> Dear @ViniciusBRodrigues, @trashbirdecology and @petroniocandido, would you be willing to review this submission for JOSS?
>
> We carry out an open checklist-driven reviews here in GitHub issues, and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> This is still a pre-review issue, and once there are enough reviewers we will move on to a new issue where the review will take place. If you are not able to review, please let us know if you have recommendations for other reviewers that could help with this submission.

@marcosvital and @gmyenni I think that I can help, but would be better to see the article proof first.
