### 2021-03-10 21824d0
Run all actions on Ubuntu 20.04 with matching RStudio repo
GitHub Actions recently upgraded ubuntu-latest to 20.04 from 18.04.
The RStudio package manager was setup for 18.04, which was hardcoded in one action but not the others.
This change hardcodes all runs onto 20.04 and updates the RStudio package manager repos to match.

Fixes #209.
