+++ .github/workflows/check-release.yaml
@@ -12,10 +12,10 @@ name: R-CMD-check-release
 
 jobs:
   R-CMD-check:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-20.04
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      CRAN_REPO: https://packagemanager.rstudio.com/all/__linux__/bionic/latest
+      CRAN_REPO: https://packagemanager.rstudio.com/all/__linux__/focal/latest
     steps:
       - uses: actions/checkout@v2
       - uses: r-lib/actions/setup-r@v1
