+++ .github/workflows/test-coverage.yaml
@@ -10,10 +10,10 @@ name: test-coverage
 
 jobs:
   test-coverage:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-20.04
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      CRAN_REPO: https://packagemanager.rstudio.com/all/__linux__/bionic/latest
+      CRAN_REPO: https://packagemanager.rstudio.com/all/__linux__/focal/latest
     steps:
       - uses: actions/checkout@v2
 
