### 2021-03-10 1890a55
Reinstall stringi
R 4.0.4 can not install stringi, thus rmarkdown, correctly
