+++ .github/workflows/pkgdown.yaml
@@ -13,6 +13,8 @@ jobs:
   pkgdown:
     runs-on: ubuntu-20.04
     env:
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      _R_CHECK_FORCE_SUGGESTS_: true
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       CRAN_REPO: https://packagemanager.rstudio.com/all/__linux__/focal/latest
     steps:
@@ -47,10 +49,13 @@ jobs:
       - name: Install package dependencies
         run: |
           remotes::install_deps(dependencies = TRUE, repos = c("CRAN" = Sys.getenv("CRAN_REPO")))
+          remove.packages("knitr")
+          remove.packages("stringi")
+          remove.packages("stringr")
+          install.packages(c("knitr", "roxygen2", "selectr", "selectr", "tidyverse", "stringi", "stringr"))
           remotes::install_github("tidyverse/tidytemplate")
           remotes::install_dev("pkgdown")
         shell: Rscript {0}
-
 
       - name: Install package
         run: R CMD INSTALL .
