### 2021-03-10 6faf4c8
Stringi install
We don't need to keep all the removes, but we still need the install stringi step
