+++ .github/workflows/pkgdown.yaml
@@ -49,10 +49,7 @@ jobs:
       - name: Install package dependencies
         run: |
           remotes::install_deps(dependencies = TRUE, repos = c("CRAN" = Sys.getenv("CRAN_REPO")))
-          remove.packages("knitr")
-          remove.packages("stringi")
-          remove.packages("stringr")
-          install.packages(c("knitr", "roxygen2", "selectr", "selectr", "tidyverse", "stringi", "stringr"))
+          install.packages(c("stringi", "stringr", "knitr"))
           remotes::install_github("tidyverse/tidytemplate")
           remotes::install_dev("pkgdown")
         shell: Rscript {0}
