+++ working.R
@@ -4,13 +4,12 @@ looks like from this that a general ricker model with the three params
 is flexible enough that a fit of the data can generate trajectories like it
 as well as others but still its a great starting point
 
-
-
+devtools::load_all()
+main <- "./testing"
 DM <- read_rodents_table(main, "DM_controls")
 covar <- read_covariates(main)
 covar
 plot(covar$moon[1:NROW(DM)], DM[,4])
-
 
 DMs <- na.omit(DM[,4])
 Nt1 <- DMs[-length(DMs)]
@@ -24,13 +23,185 @@ Rbar <- Rt2 / t2_t1
 
 hist(Rbar)
 
+quiet <- verbose <- FALSE
+arg_checks <- TRUE
+control_files <- files_control()
+control_runjags <- runjags_control(adapt = 1000, burnin = 1000, sample = 1000)
+
+lag <- NA
+data_set <- "DM_controls"
+  rodents_table <- read_rodents_table(main = main, data_set = data_set, 
+                                      arg_checks = arg_checks)
+
+  metadata <- read_metadata(main = main, control_files = control_files,
+                            arg_checks = arg_checks)
+  data_set_controls <- metadata$controls_r[[data_set]]
+  start_moon <- metadata$start_moon
+  end_moon <- metadata$end_moon
+  true_count_lead <- length(metadata$rodent_cast_moons)
+  CL <- metadata$confidence_level
+
+
+    ss <- s <- "DM"
+    messageq(paste0("   -", ss), !verbose)
+
+    moon_in <- which(rodents_table$moon >= start_moon & 
+                     rodents_table$moon <= end_moon)
+    past_moon_in <- which(rodents_table$moon < start_moon)
+    moon <- rodents_table[moon_in, "moon"] 
+    moon <- c(moon, metadata$rodent_cast_moons)
+    past_moon <- rodents_table[past_moon_in, "moon"]
+
+    ntraps <- rodents_table[moon_in, "ntraps"] 
+    ntraps[which(is.na(ntraps) == TRUE)] <- 0
+    cast_ntraps <- rep(max(ntraps), true_count_lead)
+    ntraps <- c(ntraps, cast_ntraps)
+    past_ntraps <- rodents_table[past_moon_in, "ntraps"]
+
+    species_in <- which(colnames(rodents_table) == s)
+    count <- rodents_table[moon_in, species_in]
+    if(sum(count, na.rm = TRUE) == 0){
+      next()
+    }
+    cast_count <- rep(NA, true_count_lead)
+    count <- c(count, cast_count)
+    past_count <- rodents_table[past_moon_in, species_in]
+
+    no_count <- which(is.na(past_count) == TRUE)
+    past_moon <- past_moon[-no_count]
+    past_count <- past_count[-no_count]
+    past_ntraps <- past_ntraps[-no_count]
+
+    data <- list(count = count, ntraps = ntraps, max_ntraps = max(ntraps),
+                 N = length(count),
+                 moon = moon, past_moon = past_moon, past_count = past_count,
+                 past_ntraps = past_ntraps, past_N = length(past_count))
+
+
+model <- "model {  
+
+  mean_past_count <- mean(past_count)
+  max_past_count <- max(past_count)
[…]
