### 2021-03-17 7eb1d03
fully working logistic
now needs to be properly integrated into the existing code owrkflow for portalcasting
follows jags template structure
draft in working.R
