+++ ds.R
@@ -1,417 +1,0 @@
-main <- "./testing"
-
-data_set <- "all"
-s <- "DS"
-ss <- "DS"
-
- arg_checks <- TRUE
-
-control_files <- files_control() 
-control_runjags <- runjags_control()
-lag <- NA 
-quiet <- FALSE
-verbose <- TRUE
-
-  rodents_table <- read_rodents_table(main = main, data_set = data_set, 
-                                      arg_checks = arg_checks)
-
-  metadata <- read_metadata(main = main, control_files = control_files,
-                            arg_checks = arg_checks)
-  data_set_controls <- metadata$controls_r[[data_set]]
-  start_moon <- metadata$start_moon
-  end_moon <- metadata$end_moon
-  true_count_lead <- length(metadata$rodent_cast_moons)
-  CL <- metadata$confidence_level
-
-N_past_moons <- 12
-past_moon_0 <- start_moon - N_past_moons
-
-    moon_in <- which(rodents_table$moon >= start_moon & 
-                     rodents_table$moon <= end_moon)
-    past_moon_in <- which(rodents_table$moon >= past_moon_0 &
-                          rodents_table$moon < start_moon)
-    moon <- rodents_table[moon_in, "moon"] 
-    moon <- c(moon, metadata$rodent_cast_moons)
-    past_moon <- rodents_table[past_moon_in, "moon"]
-
-    ntraps <- rodents_table[moon_in, "ntraps"] 
-    ntraps[which(is.na(ntraps) == TRUE)] <- 0
-    cast_ntraps <- rep(max(ntraps), true_count_lead)
-    ntraps <- c(ntraps, cast_ntraps)
-    past_ntraps <- rodents_table[past_moon_in, "ntraps"]
-
-    species_in <- which(colnames(rodents_table) == s)
-    count <- rodents_table[moon_in, species_in]
-    if(sum(count, na.rm = TRUE) == 0){
-      next()
-    }
-    cast_count <- rep(NA, true_count_lead)
-    count <- c(count, cast_count)
-    past_count <- rodents_table[past_moon_in, species_in]
-
-    no_count <- is.na(past_count) == TRUE
-    past_moon <- past_moon[!no_count]
-    past_count <- past_count[!no_count]
-    past_ntraps <- past_ntraps[!no_count]
-
-    past_N <- length(past_count)
-    past_diff_rate <- diff_count <- diff_time <- rep(NA, past_N - 1)
-    diff_count[1] <- past_count[2] - past_count[1]
-    diff_time[1] <- past_moon[2] - past_moon[1]
-    past_diff_rate[1] <- diff_count[1] / diff_time[1]
-    for(j in 2:(past_N - 1)){
-      diff_count[j] <- past_count[j] - past_count[j-1] 
-      diff_time[j] <- past_moon[j] - past_moon[j-1]
-      past_diff_rate[j] <- diff_count[j] / diff_time[j]
-    }
-
-
-    data <- list(count = count, ntraps = ntraps, max_ntraps = max(ntraps),
-                 N = length(count), past_diff_rate = past_diff_rate,
-                 moon = moon, past_moon = past_moon, past_count = past_count,
-                 past_ntraps = past_ntraps, past_N = past_N)
-
-
-# remove X !!!
-  monitor <-c( "mu", "r", "K", "a", "tau", "X")
-
-  inits <- function(data = NULL){
-    rngs <- c("base::Wichmann-Hill", "base::Marsaglia-Multicarry",
-              "base::Super-Duper", "base::Mersenne-Twister")
-
-    past_count <- data$past_count 
-    past_moon <- data$past_moon 
[…]
