+++ ricker_log.R
@@ -1,207 +1,0 @@
-
-#
-# trying to get this to work for the critical species
-#
-#   dm do pp ot pb 
-#   pe rm
-#
-#   big issue is the K estimation
-#   look like it works now if i just generalize the range 
-#
-
-devtools::load_all()
-main <- "./testing"
-
-data_set <- "all"
-s <- "DM"
-ss <- "DM"
-
- arg_checks <- TRUE
-
-control_files <- files_control() 
-control_runjags <- runjags_control()
-lag <- NA 
-quiet <- FALSE
-verbose <- TRUE
-
-  rodents_table <- read_rodents_table(main = main, data_set = data_set, 
-                                      arg_checks = arg_checks)
-
-  metadata <- read_metadata(main = main, control_files = control_files,
-                            arg_checks = arg_checks)
-  data_set_controls <- metadata$controls_r[[data_set]]
-  start_moon <- metadata$start_moon
-  end_moon <- metadata$end_moon
-  true_count_lead <- length(metadata$rodent_cast_moons)
-  CL <- metadata$confidence_level
-
-    moon_in <- which(rodents_table$moon >= start_moon & 
-                     rodents_table$moon <= end_moon)
-    past_moon_in <- which(rodents_table$moon < start_moon)
-    moon <- rodents_table[moon_in, "moon"] 
-    moon <- c(moon, metadata$rodent_cast_moons)
-    past_moon <- rodents_table[past_moon_in, "moon"]
-
-    ntraps <- rodents_table[moon_in, "ntraps"] 
-    ntraps[which(is.na(ntraps) == TRUE)] <- 0
-    cast_ntraps <- rep(max(ntraps), true_count_lead)
-    ntraps <- c(ntraps, cast_ntraps)
-    past_ntraps <- rodents_table[past_moon_in, "ntraps"]
-
-    species_in <- which(colnames(rodents_table) == s)
-    count <- rodents_table[moon_in, species_in]
-    if(sum(count, na.rm = TRUE) == 0){
-      next()
-    }
-    cast_count <- rep(NA, true_count_lead)
-    count <- c(count, cast_count)
-
-    past_count <- rodents_table[past_moon_in, species_in]
-    no_count <- is.na(past_count) == TRUE
-    past_moon <- past_moon[!no_count]
-    past_count <- past_count[!no_count]
-    past_ntraps <- past_ntraps[!no_count]
-
-    log_past_count <- log(apply(cbind(past_count, 1e-3), 1, max))
-
-
-    past_N <- length(past_count)
-    past_diff_log_rate <- diff_log_count <- diff_time <- rep(NA, past_N - 1)
-
-    diff_log_count[1] <- log_past_count[2] - log_past_count[1]
-    diff_time[1] <- past_moon[2] - past_moon[1]
-    past_diff_log_rate[1] <- diff_log_count[1] / diff_time[1]
-    for(j in 2:(past_N - 1)){
-      diff_log_count[j] <- log_past_count[j] - log_past_count[j-1] 
-      diff_time[j] <- past_moon[j] - past_moon[j-1]
-      past_diff_log_rate[j] <- diff_log_count[j] / diff_time[j]
-    }
-
-
-    data <- list(count = count, ntraps = ntraps, max_ntraps = max(ntraps),
-                 N = length(count), past_diff_log_rate = past_diff_log_rate,
-                 moon = moon, past_moon = past_moon, 
-                 past_ntraps = past_ntraps, past_N = past_N,
-                 log_past_count = log_past_count)
-
-
-
-  monitor <-c( "log_mu", "r", "log_K", "a", "tau", "X")
-
[…]
