+++ jags_rw.R
@@ -25,7 +25,7 @@ ss <- "DM"
 arg_checks <- TRUE
 
 control_files <- files_control() 
-control_runjags <- runjags_control()
+control_runjags <- runjags_control(nchains = 4)
 lag <- NA 
 quiet <- FALSE
 verbose <- TRUE
@@ -46,7 +46,7 @@ CL <- metadata$confidence_level
 # add to metadata
 
 start_past_moon <- 200
-min_val <- 1e-3
+min_val <- 0.1
 
 # make these into functions too
 
@@ -71,7 +71,7 @@ past_ntraps <- past_ntraps[!na_past_count]
 past_moon <- past_moon[!na_past_count]
 past_N <- length(past_count)
 
-min_sd <- 1e-3
+min_sd <- 0.01
 
 log_past_count <- log(past_count + min_val)
 mean_log_past_count <- mean(log_past_count)
@@ -104,15 +104,17 @@ data <- list(pred_log_x1 = pred_log_x1,
              T = N,       # T = time steps
              c = ntraps,  # c = cap
              max_c = max(ntraps),
-             min_val = min_val)
+             min_val = min_val,
+             log_max_c = log(max(ntraps)),
+             log_min_val = log(min_val))
 
 
 
 jags_model <- "model {  
 
-  log_x1 ~ dnorm(pred_log_x1, precision_pred_log_x1)
+  log_x1 ~ dnorm(pred_log_x1, precision_pred_log_x1) T(log_min_val, log_max_c)
 
-  tau_delta ~ dgamma(shape_delta, rate_delta)
+  tau_delta ~ dgamma(shape_delta, rate_delta) T(1e-50, )
 
   log_x[1] <- min(log_x1, log(max_c))
   x[1] <- max(c(exp(log_x[1]), min_val))
@@ -137,10 +139,12 @@ inits <- function(data = NULL, generators = rngs(), seeds = 1:1e6){
 
     list(.RNG.name = sample(generators, 1),
          .RNG.seed = sample(seeds, 1),
-         log_x1 = max(c(rnorm(1, data$pred_log_x1, data$sd_pred_log_x1),
-                        log(data$max_c))),
-         tau_delta = rgamma(1, shape = data$shape_delta, 
-                               rate = data$rate_delta)     
+         log_x1 = max(c(min(c(rnorm(1, data$pred_log_x1, data$sd_pred_log_x1),
+                            data$log_max_c)),
+                        data$log_min_val)),
+         tau_delta = max(c(rgamma(1, shape = data$shape_delta, 
+                                  rate = data$rate_delta),
+                           1e-50))     
     )
 
   }
@@ -171,12 +175,4 @@ plot(modd, vars = c("log_x1", "tau_delta"))
 
 plot(count)
 points(modd_sum[3:NROW(modd_sum), 2], type = "l")
-
-
-
-
-
-
-count
-plot(count)
 
