+++ working.R
@@ -1,722 +1,11 @@
 #
-#  working version of a logistic growth model
+#  have a series of jags models of increasing complexity all parameterized
+#   using the same terms etc
 #
-
-# #
 #
-#  generalizations occurring.
-#  see log version too
+#  can we get it to throw an error and stop when it hangs instead of hanging?
 #
-
 devtools::load_all()
 main <- "./testing"
 
-
-
-
-jags_Ricker <- function(main = ".", data_set = "all",  
-                        control_files = files_control(), 
-                        control_runjags = runjags_control(), lag = NA, 
-                        quiet = FALSE, verbose = FALSE, arg_checks = TRUE){
-
-  check_args(arg_checks = arg_checks)
-  data_set <- tolower(data_set)
-  messageq(paste0("  -jags_Ricker for ", data_set), quiet)
-  covariatesTF <- ifelse(is.na(lag), FALSE, TRUE)
-
-  monitor <-c( "mu", "r", "K", "a", "tau")
-
-  inits <- function(data = NULL){
-    rngs <- c("base::Wichmann-Hill", "base::Marsaglia-Multicarry",
-              "base::Super-Duper", "base::Mersenne-Twister")
-
-    past_count <- data$past_count 
-    past_moon <- data$past_moon 
-    past_N <- data$past_N
-    moon <- data$moon
-    max_ntraps <- data$max_ntraps
-
-    past_diff_rate <- data$past_diff_rate 
-
-    mean_past_count <- mean(past_count)
-    max_past_count <- max(past_count)
-    sd_past_count <- sd(past_count)
-    
-    last_count <- past_count[past_N]
-    last_moon <- past_moon[past_N]
-    first_moon <- moon[1]
-    last_time_diff <- first_moon - last_moon
-    
-    mean_past_diff_rate <- mean(past_diff_rate)
-    sd_past_diff_rate <- max(c(sd(past_diff_rate) * sqrt(2), 0.01))
-    var_past_diff_rate <- sd_past_diff_rate^2
-    precision_past_diff_rate <- 1/(var_past_diff_rate)
-
-    pred_mu <- last_count + last_time_diff * mean_past_diff_rate
-
-    rate <- 100
-    shape <- precision_past_diff_rate * rate
-
-    function(chain = chain){
-      list(.RNG.name = sample(rngs, 1),
-           .RNG.seed = sample(1:1e+06, 1),
-            mu = max(c(rnorm(1, pred_mu, sd_past_count), 1e-4)),
-            r = rnorm(1, mean_past_diff_rate, 
-                         max(c(1e-4, mean_past_diff_rate))),
-            K = runif(1, 2, max_past_count * 2), 
-            a = runif(1, 0.5, 1.5),
-            tau = rgamma(1, shape = shape, rate = rate))
-    }
-  }
-
-
-  jags_model <- "model {  
-
-    mean_past_count <- mean(past_count)
-    max_past_count <- max(past_count)
-    sd_past_count <- max(c(sd(past_count) * sqrt(2), 0.01))
-    var_past_count <- sd_past_count^2
-    precision_past_count <- 1/(var_past_count)
-
-    last_count <- past_count[past_N]
-    last_moon <- past_moon[past_N]
-    first_moon <- moon[1]
-    last_time_diff <- first_moon - last_moon
-
-    mean_past_diff_rate <- mean(past_diff_rate)
[…]
