### 2021-03-21 811a576
generalized ricker model (and sub models) working for all species
in a standalone space (not yet fully integrated into portalcasting function pipeline), the
mean
random walk
logistic
generalized ricker

functions work with
DM, DO, PP, OT, PB, PE, RM
