### 2021-03-31 #3027 @FlukeAndFeather
> Hi @marcosvital , unfortunately I am not able to review this in the next month. Perhaps @FlukeAndFeather is interested and available?

@pboesu thanks for suggesting me. Yes, I would be happy to review `portalcasting`. This would be my first JOSS review - is there a sign-up process? I read through the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and didn't see that mentioned. Thanks!
