+++ jags_rw.R
@@ -19,7 +19,7 @@ rngs <- function(){
 }
 
 
-data_set <- "all"
+data_set <- "controls"
 s <- "DM"
 ss <- "DM"
 arg_checks <- TRUE
