+++ jags_ricker_generalized.R
@@ -19,9 +19,9 @@ rngs <- function(){
 
 
 
-data_set <- "all"
-s <- "RM"
-ss <- "RM"
+data_set <- "controls"
+s <- "DM"
+ss <- "DM"
 arg_checks <- TRUE
 
 control_files <- files_control() 
@@ -186,9 +186,9 @@ monitor <- c("log_x1", "tau_delta", "r", "K", "a", "x")
 modd <- run.jags(model = jags_model, monitor = monitor, 
                           inits = inits(data), data = data,  
                           n.chains = control_runjags$nchains,
-                          adapt = control_runjags$adapt,
-                          burnin = control_runjags$burnin,
-                          sample = control_runjags$sample,
+                          adapt = 1e3,
+                          burnin = 1e3,
+                          sample = 1e3,
                           thin = control_runjags$thin, 
                           modules = control_runjags$modules, 
                           method = control_runjags$method, 
