### 2021-04-27 #3220 @whedon
[REVIEW]: portalcasting: Supporting automated forecasting of rodent populations
**Submitting author:** @gmyenni (<a href="http://orcid.org/0000-0001-6969-1848">Glenda Yenni</a>)
**Repository:** <a href="http://github.com/weecology/portalcasting" target ="_blank">http://github.com/weecology/portalcasting</a>
**Version:** v0.20.0
**Reviewer:** @ViniciusBRodrigues , @FlukeAndFeather
