### 2021-04-28 ea0c075
adding henry to authors list in description
was in zenodo only accidentally
