### 2021-05-07 #3027 @neerajdhanraj
Dear Marcos,

Sorry to see this email in my Spam folder. I know it must be already too
late to respond you. Are you still awaiting for reviews on this paper?

Best Regards,

*Neeraj Dhanraj Bokde     *
Postdoctoral Researcher

Ph.D., M.E.

*     Aarhus University*

*     Department of Mechanical and Production Engineering*

     Inge Lehmanns Gade 10, Navitas

     8000  Aarhus C

     Denmark

Tel.: +45 6090 1986
E: ***@***.***

On Fri, Apr 9, 2021 at 12:03 AM Marcos Vital (LEQ-UFAL) <
***@***.***> wrote:

> Dear @neerajdhanraj <https://github.com/neerajdhanraj>, would you be
> willing to review this submission for JOSS?
>
> We carry out an open checklist-driven reviews here in GitHub issues, and
> follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> This is still a pre-review issue, and once there are enough reviewers we
> will move on to a new issue where the review will take place. If you are
> not able to review, please let us know if you have recommendations for
> other reviewers that could help with this submission.
>
