### 2021-05-07 #3220 @ViniciusBRodrigues
Hi @marcosvital I thought that I have accepted the invitation here, but when I clicked in the link I received: "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account."

Could you check it please?
