### 2021-05-10 #3220 @ViniciusBRodrigues
@gmyenni can you improve the **Community guidelines** in your GitHub page (and also in https://weecology.github.io/portalcasting/)? There are not *clear guidelines* for that there. For someone which doesn't use GitHub it could be trickier.
