+++ R/portalcasting.R
@@ -4,8 +4,8 @@
 #' @importFrom grDevices grey rgb
 #' @importFrom httr content GET stop_for_status
 #' @importFrom jsonlite fromJSON serializeJSON unserializeJSON write_json
-#' @importFrom portalr get_future_moons load_trapping_data ndvi 
-#'  summarize_rodent_data weather
+#' @importFrom portalr download_observations get_future_moons 
+#'  load_trapping_data ndvi normalized_file_path summarize_rodent_data weather
 #' @importFrom runjags run.jags runjags.options
 #' @importFrom rEDM simplex tde_gp
 #' @importFrom scoringRules crps_sample
@@ -15,7 +15,7 @@
 #' @importFrom utils download.file packageDescription read.csv sessionInfo
 #'  tail unzip write.csv write.table
 #' @importFrom viridis viridis
-#' @importFrom yaml as.yaml yaml.load_file
+#' @importFrom yaml as.yaml yaml.load_file write_yaml
 
 #' @title Functions for Portalcasting (Forecasting Portal Rodents)
 #'
