+++ man/download_archive.Rd
@@ -1,0 +1,39 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/download.R
+\name{download_archive}
+\alias{download_archive}
+\title{Download the Portal Predictions Repository Archive}
+\usage{
+download_archive(
+  path = resources_path(main = "."),
+  version = "latest",
+  source = "github",
+  quiet = FALSE,
+  timeout = getOption("timeout")
+)
+}
+\arguments{
+\item{path}{\code{character} path to the folderinto which the archive will 
+be downloaded.}
+
+\item{version}{\code{character} version of the data to download. Default 
+\code{"latest"} downloads the most recent (by date 
+published).}
+
+\item{source}{\code{character} indicator of the source for the download. 
+Either \code{"github"} (defualt) or \code{"github"}.}
+
+\item{quiet}{\code{logical} indicator if progress messages should be
+quieted.}
+
+\item{timeout}{\code{numeric} value passed to \code{\link[base]{options}}
+to dictate the download timeout limit.}
+}
+\value{
+\code{NULL}, \code{\link[base]{invisible}}-ly.
+}
+\description{
+Downloads a specific \code{version} of the Portal Predictions 
+             repository from either GitHub or Zenodo (based on 
+             \code{source}).
+}

[…]
