### 2021-05-26 #3220 @FlukeAndFeather
@gmyenni Two things I want to bring to your attention. The first is pretty minor: a few errors in the manuscript's bibliography. Brown (1998) is missing editors and a few references have incorrect capitalization ("portal" when referring to Portal, AZ and "Portalr" when referring to the "portalr" pkg).

The second issue is: I'm running into errors trying to walk through the Getting Started vignette. After installing portalcasting from GitHub, I encounter the following.

```
library(portalcasting)
main <- "portalcast_directory"
setup_dir(main = main)
portalcast(main = main, models = c("ESSS", "AutoArima", "NaiveArima"))
plot_cast_ts(main = main)
# Error: cast not available for requested species
traceback()
# 2: stop("cast not available for requested species", call. = FALSE)
# 1: plot_cast_ts(main = main)
```

Is the Getting Started vignette out of date? Or is it an issue with my computational environment? If running a Docker image is essential, then (given the target audience) I think the tutorial on spinning up a container is required documentation.
