+++ use_case.R
@@ -1,0 +1,38 @@
+
+yoooooooo cut out normalized file path oop
+
+
+# setup_dir(main = "./portalcasting", portalPredictions_version = "latest")
+
+
+# we do want the user to easily be able to add to things being filled, 
+# but the bigger thing is what now
+#
+# because right, on the server, chill to go portalcast() but user building
+#  sandboxing, i want it to stop right here.
+#
+
+
+i'm creating a new model
+it will need new data available
+
+
+
+a model has a function
+a model is not a function
+
+
+model <- list(name = "aa", 
+              
+
+
+
+
+
+
+
+
+
+
+
+
