+++ scraps.R
@@ -1,0 +1,532 @@
+# functions that are being removed, but are likely useful still elsewhere
+
+
+prefab_model_controls <- function(){
+  list(
+    AutoArima = list(name = "AutoArima", 
+                     data_sets = c("all", "controls", "exclosures",
+                                   "dm_controls"),
+                     covariatesTF = FALSE, 
+                     lag = NA), 
+    ESSS = list(name = "ESSS", 
+                data_sets = c("all_interp", "controls_interp", 
+                              "exclosures_interp", "dm_controls_interp"),
+                covariatesTF = FALSE, 
+                lag = NA), 
+    NaiveArima = list(name = "NaiveArima", 
+                      data_sets = c("all", "controls", "exclosures",
+                                    "dm_controls"),
+                      covariatesTF = FALSE, 
+                      lag = NA), 
+    nbGARCH = list(name = "nbGARCH", 
+                   data_sets = c("all_interp", "controls_interp", 
+                                 "exclosures_interp", "dm_controls_interp"),
+                   covariatesTF = FALSE, 
+                   lag = NA), 
+    nbsGARCH = list(name = "nbsGARCH", 
+                    data_sets = c("all_interp", "controls_interp", 
+                                  "exclosures_interp", "dm_controls_interp"),
+                    covariatesTF = FALSE, 
+                    lag = NA), 
+    pevGARCH = list(name = "pevGARCH",  
+                    data_sets = c("all_interp", "controls_interp", 
+                                  "exclosures_interp", "dm_controls_interp"),
+                    covariatesTF = TRUE, lag = 6), 
+    simplexEDM = list(name = "simplexEDM", 
+                      data_sets = c("all_interp", "controls_interp", 
+                                    "exclosures_interp", 
+                                     "dm_controls_interp"), 
+                      covariatesTF = FALSE, lag = NA, max_E = 7),
+    GPEDM = list(name = "GPEDM", 
+                 data_sets = c("all_interp", "controls_interp", 
+                               "exclosures_interp", "dm_controls_interp"), 
+                 covariatesTF = FALSE, lag = NA, max_E = 7),
+    
+    jags_RW = list(name = "jags_RW", 
+                   data_sets = c("all", "controls", "exclosures",
+                                 "dm_controls"), 
+                   covariatesTF = FALSE, lag = NA,
+                   control_runjags = runjags_control())
+  )
+}
+
+
+#' @title Produce the control lists for models
+#' 
+#' @description The models are written and managed using generalized functions
+#'  that take a few arguments that vary among models, and so this function
+#'  produces the appropriate \code{list} of \code{list}s of model
+#'  script-writing control arguments. The specific \code{models} must be
+#'  input by the user, with the default being \code{NULL}, which returns.
+#'  \code{NULL}. The prefab models (\code{\link{prefab_models}}) have their
+#'  control lists included already, but any user-created model needs to
+#'  have controls provided. See \code{Details}. \cr 
+#'  \strong{Users adding models to the prefab set should add their
+#'  script-writing controls lists to the \code{prefab_controls} object in 
+#'  the source code of the function.} \cr \cr
+#'  \code{extract_min_lag} provides a simple function to extract the 
+#'  minimum non-0 lag time used across models from the controls lists.
+#'
+#' @details Any model that is part of the \code{prefab} set 
[…]
