+++ R/utilities.R
@@ -1,3 +1,230 @@
+#' @title Optionally generate a message based on a logical input
+#'
+#' @description A wrapper on \code{\link[base]{message}} that, given the
+#'              input to \code{quiet}, generates the message(s) in \code{...}
+#'              or not.
+#'
+#' @param ... zero or more objects that can be coerced to \code{character}
+#'            and are concatenated with no separator added, or a single 
+#'            condition object.
+#'            \cr \cr
+#'            See \code{\link[base]{message}}.
+#'
+#' @param quiet \code{logical} indicator if the message should be generated. 
+#'
+#' @param domain The domain for the translation. If \code{NA}, messages will
+#'               not be translated.
+#'               \cr \cr
+#'               See \code{\link[base]{message}} and 
+#'               \code{\link[base]{gettext}}.
+#'
+#' @param appendLF \code{logical} indicator if messages given as a 
+#'                 \code{character} string should have a newline appended.
+#'                 \cr \cr
+#'                 See \code{\link[base]{message}}.
+#'
+#' @return A message is given, and \code{NULL} returned.
+#'
+#' @export
+#'
+messageq <- function (..., 
+                      quiet    = FALSE, 
+                      domain   = NULL, 
+                      appendLF = TRUE) {
+
+  if (!quiet) {
+
+    message(...,
+            domain   = domain,
+            appendLF = appendLF)
+
+  }
+
+  invisible()
+}
+
+
+
+#' @title Determine the path for a specific level of a portalcasting directory
+#'
+#' @description Produce paths for a portalcasting directory. 
+#'              \cr \cr
+#'              \code{main_path} returns the path for the \code{main} folder.
+#'              \cr \cr
+#'              \code{sub_path} returns the path(s) for the \code{subs}
+#'              folder(s). \cr \cr
+#'
+#' @param main \code{character} value of the name of the main component of
+#'              the directory tree. 
+#'
+#' @param sub \code{character} vector of the name(s) of the sub component(s)
+#'             of the directory tree. 
+#'
+#' @return \code{character} \code{\link[port{normalized_file_path}}(s).
+#'          \cr \cr
+#'          \code{main_path} path of the \code{main} folder.
+#'          \cr \cr
+#'          \code{subs_path} paths of the \code{sub} folder(s). 
+#'
+#' @name paths
+#'
+NULL
+
+#' @rdname paths
+#'
+#' @export
+#'
+main_path <- function (main = ".") {
+
+  if (length(main) > 1) {
+
+    stop("`main` must be length-1")
+
+  }
+
+  normalized_file_path(main,
+                       mustWork = FALSE)
+
+}
+
+#' @rdname paths
+#'
+#' @export
+#'
+sub_path <- function (main = ".", sub = NULL) {
+
+  if (length(main) > 1) {
+
+    stop("`main` must be length-1")
+
+  }
+
+  normalized_file_path(main, sub,
+                       mustWork = FALSE)
+
+}
+
+#' @rdname paths
+#'
+#' @export
+#'
+resources_path <- function (main = ".") {
+
+  sub_path(main = main, 
+           sub = "resources")
+
+}
+
+#' @rdname paths
+#'
+#' @export
+#'
+casts_path <- function (main = ".") {
+
+  sub_path(main = main, 
+           sub = "casts")
+
+}
+
+#' @rdname paths
+#'
+#' @export
+#'
+fits_path <- function (main = ".") {
+
+  sub_path(main = main, 
+           sub = "fits")
+
+}
+
+#' @rdname paths
+#'
+#' @export
+#'
[…]

[…]
