### 2021-07-04 45f4e51
pulled everything not integrated out to higher level
trying to avoid vestigials etc
