+++ R/prepare_covariates.R
@@ -1,334 +1,0 @@
-#' @title Prepare covariate data for casting
-#'
-#' @description Combine the historical and cast covariate data for a model
-#'  run. \cr \cr See \code{\link{prep_hist_covariates}} and
-#'  \code{\link{prep_cast_covariates}} for the historical and -casted
-#'  covariate preparation details, respectively.
-#'
-#' @param main \code{character} value of the name of the main component of
-#'  the directory tree.
-#'
-#' @param moons Moons \code{data.frame}. See \code{\link{prep_moons}}.
-#'
-#' @param lead_time \code{integer} (or integer \code{numeric}) value for the
-#'  number of timesteps forward a cast will cover.
-#'
-#' @param min_lag \code{integer} (or integer \code{numeric}) of the minimum 
-#'  covariate lag time used in any model.
-#'
-#' @param cast_date \code{Date} from which future is defined (the origin of
-#'  the cast). In the recurring forecasting, is set to today's date
-#'  using \code{\link{Sys.Date}}.
-#'
-#' @param end_moon \code{integer} (or integer \code{numeric}) newmoon number 
-#'  of the last sample to be included. Default value is \code{NULL}, which 
-#'  equates to the most recently included sample. 
-#'
-#' @param quiet \code{logical} indicator if progress messages should be
-#'  quieted.
-#'
-#' @param verbose \code{logical} indicator if detailed messages should be
-#'  shown.
-#'
-#' @param start_moon \code{integer} (or integer \code{numeric}) newmoon number 
-#'  of the first sample to be included. Default value is \code{217}, 
-#'  corresponding to \code{1995-01-01}.
-#'
-#' @param control_climate_dl \code{list} of specifications for the download, 
-#'  which are sent to \code{\link{NMME_urls}} to create the specific URLs. See
-#'  \code{\link{climate_dl_control}}.
-#'
-#' @param control_files \code{list} of names of the folders and files within
-#'  the sub directories and saving strategies (save, overwrite, append, etc.).
-#'  Generally shouldn't need to be edited. See \code{\link{files_control}}.
-#'
-#' @param arg_checks \code{logical} value of if the arguments should be
-#'  checked using standard protocols via \code{\link{check_args}}. The 
-#'  default (\code{arg_checks = TRUE}) ensures that all inputs are 
-#'  formatted correctly and provides directed error messages if not.
-#'
-#' @return \code{data.frame} of historical and -casted covariates, combined
-#'  and saved out to \code{filename} if indicated by \code{save}.
-#'  
-#' @examples
-#'  \donttest{
-#'   setup_dir()
-#'   prep_covariates()
-#'  }
-#'  
-#' @export
-#'
-prep_covariates <- function(main = ".", moons = NULL, end_moon = NULL, 
-                            start_moon = 217, lead_time = 12, min_lag = 6, 
-                            cast_date = Sys.Date(),
-                            control_climate_dl = climate_dl_control(), 
-                            control_files = files_control(),
-                            quiet = TRUE, verbose = FALSE, 
-                            arg_checks = TRUE){
-  check_args(arg_checks = arg_checks)
-  moons <- ifnull(moons, read_moons(main = main, 
-                                    control_files = control_files,
-                                    arg_checks = arg_checks))
-  messageq("  -covariate data files", quiet)
-
[…]
