+++ notes to me rn.R
@@ -1,0 +1,16 @@
+working:
+
+the models are now simply written as functions within scripts, which is ideal
+
+but need to update the models themselves t reflect how the function will be used
+very very streamlined to a single data stream
+
+
+
+the data sets are nicely organized and generated now
+can pop back to the models
+
+starting with a single simple model
+
+what should the AutoArima function accomplish?
+

[…]
