+++ paper/paper.bib
@@ -113,10 +113,12 @@ year = {2016}
  year = {1977}
 }
 
-@article{brown1998desert,
+@incollection{brown1998desert,
  author = {Brown, JH},
- journal = {Experimental ecology. Oxford University Press, Oxford, UK},
- pages = {71--95},
  title = {The desert granivory experiments at Portal},
+ editor = {W. J. Resetarits and J. Bernardo},
+ booktitle = {Experimental ecology: issues and perspectives},
+ publisher = {Oxford University Press, Oxford, UK},
+ pages = {71--95},
  year = {1998}
 }
