+++ paper/paper.bib
@@ -114,7 +114,7 @@ year = {2016}
 }
 
 @incollection{brown1998desert,
- author = {Brown, JH},
+ author = {Brown, James H},
  title = {The desert granivory experiments at Portal},
  editor = {W. J. Resetarits and J. Bernardo},
  booktitle = {Experimental ecology: issues and perspectives},
