### 2021-09-08 b587b14
Force capitalization in titles
