+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @Article{christensen2019portalr,
-    title = {portalr: an R package for summarizing and using the Portal
-      Project Data},
+    title = {{portalr}: an {R} package for summarizing and using the {Portal
+      Project Data}},
     author = {Erica M. Christensen and Glenda M. Yenni and Hao Ye and
       Juniper L. Simonis and Ellen K. Bledsoe and Renata M. Diaz and
       Shawn D. Taylor and Ethan P. White and S. K. Morgan Ernest},
@@ -14,7 +14,7 @@
   
 @article{ernest2009data,
 author = {Ernest, S. K. Morgan and Valone, Thomas J. and Brown, James H.},
-title = {Long-term monitoring and experimental manipulation of a Chihuahuan Desert ecosystem near Portal, Arizona, USA},
+title = {Long-term monitoring and experimental manipulation of a {Chihuahuan} desert ecosystem near {Portal, Arizona, USA}},
 journal = {Ecology},
 volume = {90},
 number = {6},
@@ -29,7 +29,7 @@ year = {2009}
 
 @article{ernest2016data,
 author = {Morgan Ernest, S. K. and Yenni, Glenda M. and Allington, Ginger and Christensen, Erica M. and Geluso, Keith and Goheen, Jacob R. and Schutzenhofer, Michelle R. and Supp, Sarah R. and Thibault, Katherine M. and Brown, James H. and Valone, Thomas J.},
-title = {Long-term monitoring and experimental manipulation of a Chihuahuan desert ecosystem near Portal, Arizona (1977–2013)},
+title = {Long-term monitoring and experimental manipulation of a {Chihuahuan} desert ecosystem near {Portal, Arizona} (1977–2013)},
 journal = {Ecology},
 volume = {97},
 number = {4},
@@ -62,7 +62,7 @@ year = {2016}
     
 @article{Ernest:2018b,
     author = {S. K. Morgan {Ernest} and Glenda Marie {Yenni} and Ginger {Allington} and Ellen {Bledsoe} and Erica {Christensen} and Renata {Diaz} and Jacob {Goheen} and Qinfeng {Guo} and Edward {Heske} and Douglas {Kelt} and Joan {Meiners} and Jim {Munger} and Carla {Restrepo} and Douglas {Samson} and Michele {Schutzenhofer} and Marian {Skupski} and Sarah {Supp} and Katherine {Thibault} and Shawn {Taylor} and Ethan {White} and Diane {Davidson} and James H {Brown} and Thomas J {Valone}},
-    title={The Portal Project: a long-term study of a Chihuahuan desert ecosystem},
+    title={The {Portal Project}: a long-term study of a {Chihuahuan} desert ecosystem},
     url={https://www.biorxiv.org/content/early/2018/05/28/332783},
     doi={10.1101/332783},
     year={2018},
@@ -115,7 +115,7 @@ year = {2016}
 
 @incollection{brown1998desert,
  author = {Brown, James H},
- title = {The desert granivory experiments at Portal},
+ title = {The desert granivory experiments at {Portal}},
  editor = {W. J. Resetarits and J. Bernardo},
  booktitle = {Experimental ecology: issues and perspectives},
  publisher = {Oxford University Press, Oxford, UK},
