+++ pkgdown.yml
@@ -1,13 +1,13 @@
 pandoc: 2.7.3
-pkgdown: 1.6.1.9000
-pkgdown_sha: d0a93b68f9eaf951ba27c5ec74175acfdbe8b2de
+pkgdown: 1.6.1.9001
+pkgdown_sha: 705ff7c650bb1c7d46d35e72f27ad093689e2f29
 articles:
   adding_model_and_data: adding_model_and_data.html
   codebase: codebase.html
   current_models: current_models.html
   getting_started: getting_started.html
   jags_models: jags_models.html
-last_built: 2021-03-10T23:25Z
+last_built: 2021-09-08T17:48Z
 urls:
   reference: https://weecology.github.io/portalcasting/reference
   article: https://weecology.github.io/portalcasting/articles
