+++ paper/paper.bib
@@ -110,7 +110,9 @@ year = {2016}
  publisher = {American Association for the Advancement of Science},
  title = {Competition between seed-eating rodents and ants in desert ecosystems},
  volume = {196},
- year = {1977}
+ year = {1977},
+ doi = {10.1126/science.196.4292.880}
+
 }
 
 @incollection{brown1998desert,
