+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   current_models: current_models.html
   getting_started: getting_started.html
   jags_models: jags_models.html
-last_built: 2021-09-08T17:48Z
+last_built: 2021-09-08T17:55Z
 urls:
   reference: https://weecology.github.io/portalcasting/reference
   article: https://weecology.github.io/portalcasting/articles
