+++ DESCRIPTION
@@ -42,6 +42,7 @@ Imports:
 Suggests:
   knitr,
   pkgdown,
+  markdown,
   rmarkdown (>= 1.16.1),
   testthat
 RoxygenNote: 7.1.1
