### 2021-09-16 e2e36fe
Add docker build and push workflow
