+++ .github/workflows/docker-publish.yml
@@ -58,6 +58,6 @@ jobs:
         with:
           context: .
           push: ${{ github.event_name != 'pull_request' }}
-          tags: ${{ steps.meta.outputs.tags }}
+          tags: ${{ steps.meta.outputs.tags }},latest 
           labels: ${{ steps.meta.outputs.labels }}
           file: docker/Dockerfile
