### 2021-09-16 14e02c6
Update docker-publish.yml
Push latest image
