+++ .github/workflows/docker-publish.yml
@@ -58,6 +58,7 @@ jobs:
         with:
           context: .
           push: ${{ github.event_name != 'pull_request' }}
-          tags: ${{ steps.meta.outputs.tags }},latest 
+          # tags: ${{ steps.meta.outputs.tags }} 
+          tags: weecology/portalcasting:latest
           labels: ${{ steps.meta.outputs.labels }}
           file: docker/Dockerfile
