### 2021-09-16 9ea2c3d
Update docker-publish.yml
Test if latest will build with other tags
