+++ .github/workflows/docker-publish.yml
@@ -47,10 +47,17 @@ jobs:
       # https://github.com/docker/metadata-action
       - name: Extract Docker metadata
         id: meta
-        uses: docker/metadata-action@98669ae865ea3cffbcbaa878cf57c20bbf1c6c38
-        with:
-          images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
-
+        run: |
+          if [[ $GITHUB_REF == 'refs/tags/'* ]]; then
+            TAGS='["weecology/portalcasting:latest","weecology/portalcasting:'${GITHUB_REF/refs\/tags\//}'"]'
+          else
+            TAGS='["weecology/portalcasting:latest"]'
+          fi
+          echo '::set-output name=tags::'$TAGS
+#         uses: docker/metadata-action@98669ae865ea3cffbcbaa878cf57c20bbf1c6c38
+#         with:
+#           images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
+          
       # Build and push Docker image with Buildx (don't push on PR)
       # https://github.com/docker/build-push-action
       - name: Build and push Docker image
@@ -58,7 +65,7 @@ jobs:
         with:
           context: .
           push: ${{ github.event_name != 'pull_request' }}
-          # tags: ${{ steps.meta.outputs.tags }} 
-          tags: weecology/portalcasting:latest
+          tags: ${{ steps.meta.outputs.tags }} 
+#           tags: weecology/portalcasting:latest
           labels: ${{ steps.meta.outputs.labels }}
           file: docker/Dockerfile
