### 2021-09-16 f311dc1
Update docker-publish.yml
Test with crazy max metadata action
