+++ .github/workflows/docker-publish.yml
@@ -45,19 +45,15 @@ jobs:
 
       # Extract metadata (tags, labels) for Docker
       # https://github.com/docker/metadata-action
-      - name: Extract Docker metadata
+      - name: Docker meta
         id: meta
-        run: |
-          if [[ $GITHUB_REF == 'refs/tags/'* ]]; then
-            TAGS='["weecology/portalcasting:latest","weecology/portalcasting:'${GITHUB_REF/refs\/tags\//}'"]'
-          else
-            TAGS='["weecology/portalcasting:latest"]'
-          fi
-          echo '::set-output name=tags::'$TAGS
-#         uses: docker/metadata-action@98669ae865ea3cffbcbaa878cf57c20bbf1c6c38
-#         with:
-#           images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
-          
+        uses: crazy-max/ghaction-docker-meta@v1
+        with:
+          images: weecology/portalcasting
+          tag-custom: latest
+          tag-semver: |
+            {{raw}}   
+
       # Build and push Docker image with Buildx (don't push on PR)
       # https://github.com/docker/build-push-action
       - name: Build and push Docker image
