+++ .travis.yml
@@ -1,37 +1,0 @@
-language: r
-cache: packages
-sudo: required
-warnings_are_errors: false
-
-before_install:
-  - sudo apt-get update -qq
-  - sudo apt-get install texlive-latex-base libharfbuzz-dev libfribidi-dev
-  - sudo apt-get install gsl-bin libgsl0-dev
-
-r_binary_packages:
-  - runjags
-
-r_packages:
-  - covr
-
-script: 
-  - |
-    R CMD build .
-    travis_wait 50 R CMD check portalcasting*tar.gz
-
-matrix:
-  include:
-  - r: devel
-  - r: release
-    after_success:
-    - R CMD INSTALL .
-    - Rscript -e 'pkgdown::build_site(examples = FALSE)'
-    - travis_wait 50 Rscript -e 'covr::codecov()'
-    deploy:
-      provider: pages
-      skip-cleanup: true
-      github-token: $GITHUB_PAT
-      keep-history: true
-      local-dir: docs
-      on:
-        branch: master
