### 2021-09-16 eaabeed
Update docker-publish.yml
Run only when there are tags
