+++ .github/workflows/check-release.yaml
@@ -19,6 +19,8 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: r-lib/actions/setup-r@v1
+      - uses: r-lib/actions/setup-pandoc@v1
+      
       - name: Install system dependencies
         run: |
           sudo apt-get update -qq
